moved enums to core
