+++ api/core/use_case/create_application_use_case.py
@@ -14,7 +14,7 @@ import pathlib
 import json
 import os
 
-from utils.enums import DMT
+from core.enums import DMT
 from core.use_case.utils.get_storage_recipe import get_storage_recipe
 from core.use_case.utils.get_template import get_blueprint
