+++ api/core/use_case/create_data_source_use_case.py
@@ -1,5 +1,5 @@
 from core.shared import request_object, response_object, use_case
-from utils.enums import DataSourceType
+from core.enums import DataSourceType
 
 
 class CreateDataSourceRequestObject(request_object.ValidRequestObject):
