+++ api/tests_bdd/steps/domain.py
@@ -9,7 +9,7 @@ from core.repository.interface.package_repository import PackageRepository
 from core.repository.mongo.blueprint_repository import MongoBlueprintRepository
 from anytree import NodeMixin, RenderTree
 from core.repository.repository_factory import get_repository, RepositoryType
-from utils.enums import SIMOS, DMT
+from core.enums import SIMOS, DMT
 
 
 class Reference:
